Move generic functions, variables and types to common.go file
