+++ geo/geopoint_test.go
@@ -264,25 +264,3 @@ func TestCircleToPolygon(t *testing.T) {
 		})
 	})
 }
-
-func TestToLowest(t *testing.T) {
-
-	Convey("Given 2 variables of type integer", t, func() {
-
-		Convey("When passing the larger integer into toLowest first", func() {
-			lowestValue := toLowest(30, 20)
-
-			Convey("Then the return of function returns the lower integer", func() {
-				So(lowestValue, ShouldEqual, 20)
-			})
-		})
-
-		Convey("When passing the smaller integer into toLowest first", func() {
-			lowestValue := toLowest(10, 40)
-
-			Convey("Then the return of function returns the lower integer", func() {
-				So(lowestValue, ShouldEqual, 10)
-			})
-		})
-	})
-}
